ReWire up Snippets collection
These files need to live in `examples` or similar to be picked up by
Antora. I've renamed them as per the previously existing files to
avoid having to update docs in multiple places.

It is not ideal for the codesnippet_collection.* files to be committed.
However:

 - there is currently no facility for Antora to run prerequisite shell
   scripts before `generate` task

 - complicating the build process requires more support from resources
   that are *in the process of* being lined up (OpenDevise, and DevOps
   contractor).

Future Approach
===============

Approach will consist of some combination of:

 * generate the snippets on build only

 * use Github Actions to ensure that collections are automatically built on
   checking in changes to snippets.
   (Also deny checkin if collection committed by itself)

 * De-abstract by removing e.g. `block_tabbed_code.adoc` and friends,
   and instead just including some code...

 * Add a "recursive include" macro that will find a tag in a directory
   path

RISK
====

Until then, someone may again try to edit the codesnippet_collections
directly. This will break things and annoy the devs.