DOC-11873: Relnote update
Release notes for CBL are now split into CBL specific and Vector Search specific release notes per platform. This is to improve visibility of both. Especially as the core CBL product is likely to be updated more regularly.

This has been reflected in the nav, the quickstart skeleton and anywhere you'd expect to find release notes.

As of yet there is no direct linking from CBL to Vector Search release notes but it's easy to find them.

Decision from PM to retain beta release notes as they were public releases.