DOC-13056 mobile simplification: p2p pages
I've verified with my human eyeballs that the output for Android pages
matches the published output, in terms of structure / images rendering / etc.

Simplification has the following known issues, due to the methodology
(using antora assembler as if to generate a single PDF)

* Some document titles are changed
(because assembler uses the left-hand nav as the authoritative source)

* Some headings are promoted
This is actually a "correction" - e.g. the original doc may have had
unsemantic headings, with a skipped level. Assembler normalises
the output and rewrites headings based on nav level.
Though we then remove the correction for nav level, we don't
"uncorrect" the fix.

* Paragraph IDs are namespaced to make them unique.
This should be *fine* within a single document, as assembler should
update both anchor and link.
But deep links from another page may be broken.

* Code samples are *inlined* rather than retained as an include::
I've tried to add a `// include::...` comment in the source so it's
easy to fix this.

In addition, you may want to request changes such as:

* back off from inlining certain {page-...} attributes to retain
genericity where actually needed.
(Probably not here, mostly an issue for e.g. installation pages)

* Any other demangling I missed.